+++ beam/www/beam/store.ts
@@ -6,6 +6,7 @@ import { computed, ref, watch } from 'vue'
 import { type RouteLocationNormalized, useRoute } from 'vue-router'
 
 import type {
+	BeamHome,
 	FormContext,
 	JobCard,
 	ListContext,
@@ -40,6 +41,14 @@ export const useDataStore = defineStore('data', () => {
 		await getScanDoctypes()
 		await setForm(currentRoute || route)
 		await setScanContext(currentRoute || route)
+	}
+
+	const getScanDoctypes = async (params?: Record<string, any>) => {
+		const url = '/api/method/beam.beam.scan.config.get_scan_doctypes'
+		const response = await get(url, params)
+		const { message }: { message: ScanConfig } = await response.json()
+		config.value = message
+		return { data: message }
 	}
 
 	// TODO: somehow vue-router's composables are not working as intended here, so accepting route input
@@ -119,11 +128,10 @@ export const useDataStore = defineStore('data', () => {
 		return data
 	}
 
-	const getScanDoctypes = async (params?: Record<string, any>) => {
-		const url = '/api/method/beam.beam.scan.config.get_scan_doctypes'
+	const getHome = async (params?: Record<string, any>) => {
+		const url = '/api/method/beam.beam.doctype.beam_settings.beam_settings.get_beam_home'
 		const response = await get(url, params)
-		const { message } = await response.json()
-		config.value = message
+		const { message }: { message: BeamHome } = await response.json()
 		return { data: message }
 	}
 
@@ -207,13 +215,6 @@ export const useDataStore = defineStore('data', () => {
 		window.location.href = '/login?redirect-to=/beam#'
 	}
 
-	const getHome = async (params?: Record<string, any>) => {
-		const url = '/api/method/beam.beam.doctype.beam_settings.beam_settings.get_beam_home'
-		const response = await get(url, params)
-		const { message } = await response.json()
-		return { data: message }
-	}
-
 	return {
 		// state
 		config,
@@ -240,11 +241,11 @@ export const useDataStore = defineStore('data', () => {
 		// other api actions
 		getAll,
 		getDemand,
-		getReceiving,
+		getHome,
 		getMappedStockEntry,
 		getOne,
-		scan,
+		getReceiving,
 		logout,
-		getHome,
+		scan,
 	}
 })
